+++ pace/paceobjects/envelopes/roof.py
@@ -1,20 +1,21 @@
 from energyreport.classes.orientation import Orientation
 
-from .pace_object import PaceObject
+from ..pace_object import PaceObject
 
 
-class Wall(PaceObject):
-    template_filename = 'templates/wallplane.xml'
-    path = './building/skin/wallPlanes/INITIAL'
+class Roof(PaceObject):
+    template_filename = 'templates/envelopes/roofplane.xml'
+    path = './building/skin/roofPlanes/INITIAL'
 
-    def __init__(self, name: str, orientation: Orientation, width: float, height: float):
+    def __init__(self, name: str, orientation: Orientation, angle: float, width: float, height: float):
         self.name = name
         self.orientation = orientation
+        self.angle = angle
         self.width = width
         self.height = height
 
     @property
-    def replace_queries(self) -> dict:
+    def replace_queries(self):
         return {
             'shortDescription': {
                 'value': self.name
@@ -22,6 +23,10 @@ class Wall(PaceObject):
             'orientation/INITIAL': {
                 'class': 'com.hemmis.mrw.pace.model.enums.Orientation',
                 'value': self.orientation.name
+            },
+            'slope/INITIAL': {
+                'class': 'java.math.BigDecimal',
+                'value': self.angle
             },
             'width/INITIAL': {
                 'class': 'java.math.BigDecimal',
